Add files via upload
simple binary search. User inputs the name and phone number. If that name or phone number is not in the phone.txt file, it doesn't return anything. If the name is searched and matches, it returns the phone number. If the phone number is searched and matches, it returns the name